% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{bar_style}
\alias{bar_style}
\title{Create a style for a bar in the vhydeg annotation table}
\usage{
bar_style(
  width = 1,
  fill = "#e6e6e6",
  height = "70\%",
  align = c("left", "right"),
  color = NULL
)
}
\arguments{
\item{width}{with of the bar}

\item{fill}{color of the bar}

\item{height}{height of the bar}

\item{align}{alignment of the bar, either "left" or "right"}

\item{color}{color of the text, if NULL, the color of the text will be the same as the color of the bar}
}
\description{
Create a style for a bar in the vhydeg annotation table
}
