% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{filter_tblastx}
\alias{filter_tblastx}
\title{Filter tblastx results to only include the AMG hit and the corresponding flanking regions}
\usage{
filter_tblastx(
  fin,
  tarvir_id,
  hit_start,
  hit_end,
  flank_len,
  remove_seq_version = T,
  ...
)
}
\arguments{
\item{fin}{Path to the tblastx file}

\item{tarvir_id}{ID of the target virus (without the AMG hit)}

\item{hit_start}{Start position of AMG hit on the query virus}

\item{hit_end}{End position of the AMG hit on the query virus}

\item{flank_len}{Length of the flanking region}

\item{remove_seq_version}{Remove the version number of the sequence ID. Default is TRUE.}

\item{...}{Other arguments}
}
\value{
data.frame
}
\description{
Filter tblastx results to only include the AMG hit and the corresponding flanking regions
}
