% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_gene2family}
\alias{get_gene2family}
\title{Read Phame gene2family annotation}
\usage{
get_gene2family(gene2family_raw)
}
\arguments{
\item{gene2family_raw}{gene2family raw data}
}
\description{
Read Phame gene2family annotation
}
