% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{read_bakta}
\alias{read_bakta}
\title{Read bakta (UniProt) annotation}
\usage{
read_bakta(fpath_bakta)
}
\arguments{
\item{fpath_bakta}{Path to bakta annotation}
}
\description{
Read bakta (UniProt) annotation
}
