% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{read_ipr}
\alias{read_ipr}
\title{Read InterProScan annotation}
\usage{
read_ipr(fpath_ipr)
}
\arguments{
\item{fpath_ipr}{Path to InterProScan annotation}
}
\description{
Read InterProScan annotation
}
