% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhydeg-package.R
\docType{package}
\name{vhydeg-package}
\alias{vhydeg}
\alias{vhydeg-package}
\title{vhydeg: Hydrocarbon degradation genes in viruses}
\description{
This is the a package with a shiny APP of viral-encoded hydrocarbon degradation genes. It contains source code for the vHYDEG project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rujinlong/vhydeg}
  \item Report bugs at \url{https://github.com/rujinlong/vhydeg/issues}
}

}
\author{
\strong{Maintainer}: Jinlong Ru \email{jinlong.ru@gmail.com}

}
\keyword{internal}
