% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vhydeg_db}
\alias{vhydeg_db}
\title{vhydeg database}
\format{
list of data.frame
}
\source{
{workflow/01_vhydeg_db.qmd}
data(vhydeg_db)
}
\usage{
vhydeg_db
}
\description{
vhydeg database
}
\keyword{datasets}
